% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{add}
\alias{add}
\title{add}
\usage{
add(l, name, x)
}
\arguments{
\item{l}{list. list to add to, *** it must have named elements ***}

\item{name}{string. Name of the new element}

\item{x.}{Element to be added (can also be a list)}
}
\value{
new list with added item
}
\description{
adds elements to lists
}
\details{
adds elements to lists at the end. If a list is passed to be added it is not flattened (i.e. it remains a list
by its own)
}
\examples{
add(list(a=1,b=2),'sublist',list(3,4))
}
\author{
Alex Zabeo
}

