% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{decimals}
\alias{decimals}
\title{decimals}
\usage{
decimals(x, k)
}
\arguments{
\item{x}{numeric. Number to be rounded}

\item{k}{numeric. Decimal positions}
}
\value{
string of rounded x with k decimals
}
\description{
rounds number to a specific decimal and converts to string
}
\details{
rounds number to a specific decimal and converts to string
}
\examples{
decimals(84.05041,2)
}
\author{
Alex Zabeo
}

