% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/main.R
\name{extrapolation}
\alias{extrapolation}
\title{extrapolation}
\usage{
extrapolation(bmd.file.name = "example.data/bmd.csv",
  efs.file.name = "example.data/Efs.csv", above.threshold = 0)
}
\arguments{
\item{bmd.file.name}{string. the location of the input file for bmds (al least it
has to have 'bmd' and 'id' columns)}

\item{efs.file.name}{string. table of EFs to be applied (see excel file in inst for instructions)}

\item{above.threshold}{numeric. If present values less equal to it are discarded from generated
bmd distribution}
}
\value{
list of two tables, the updated bmds table with results and the EFs table used
}
\description{
applies EF factors to BMD with specified distributions
}
\details{
from the BMD values extracts GM and GSD, randomly picks the same amount from the corresponding
lognormal distriution, then for each row in the EF table picks the corresponding random values
and applies the corresponding operator in cascade
}
\examples{
extrapolation("example.data/bmd.csv","example.data/Efs.csv",0)
}
\author{
Alex Zabeo
}

