% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{geomean}
\alias{geomean}
\title{geomean}
\usage{
geomean(x, na.rm = TRUE)
}
\arguments{
\item{x}{array. numbers to which calculate the geom mean}

\item{na.rm}{Bool. remove na or not}
}
\value{
geometric mean
}
\description{
geometric mean
}
\details{
evaluates geometric mean
}
\examples{
geomean(x)
}
\author{
Alex Zabeo
}

