% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{geosd}
\alias{geosd}
\title{geosd}
\usage{
geosd(x, na.rm = FALSE)
}
\arguments{
\item{x}{array. numbers to which calculate the geom sd}

\item{na.rm}{Bool. remove na or not}
}
\value{
geometric standard deviaiton
}
\description{
geometric standard deviaiton
}
\details{
evaluates geometric standard deviaiton
}
\examples{
geosd(x)
}
\author{
Alex Zabeo
}

