% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{loadData}
\usage{
loadData(file.name = "inst/data.csv", ids = c("id"), var.name = NULL)
}
\arguments{
\item{file.name}{string. name of csv file to load}

\item{ids}{vector. vector of column names to be used as ids}

\item{var.name}{string. name of the variable where to load the data in the main workspace}
}
\value{
nothing (creates variables with provided names)
}
\description{
loads necessary data from files
}
\details{
loads necessary data from files
}
\examples{
loadData("inst/data.csv",c('id'),'dat')
}
\author{
Alex Zabeo
}

