% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotall.R
\name{plotall}
\alias{plotall}
\title{plot.all}
\usage{
plotall(r, ced, iced, imoe, iexp, title, exp.type = "norm", bounds = c(-5,
  15), length = 10000, noael = NULL)
}
\arguments{
\item{r}{list. result of extrapolation()}

\item{exp.type}{c(norm,fixed). type of distribution for exposure, normal or fixed value}

\item{bmd}{string. Name of the column containing the original BMD bootstrap data}

\item{eff}{string. Name of the column containing the calculated effect threshold}

\item{exp}{string. Name of the column containing the calculated exposure value}
}
\value{
nothing, just plots
}
\description{
plot all together
}
\details{
plots bmd, effect and exposure all together in log scale
}
\examples{
plot.all(r,bmd,eff,exp)
}
\author{
Alex Zabeo
}

