% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{rpm_to_rcf}
\alias{rpm_to_rcf}
\title{rpm_to_rcf}
\usage{
rpm_to_rcf(rpm, const = 1.1179, dist = 117.4781)
}
\arguments{
\item{rpm}{numeric.}

\item{const}{numeric. Conversion constant}

\item{dist}{numeric. Distance from top of couvette}
}
\value{
rpf
}
\description{
converts rpm to rcf
}
\details{
converts rpm to rcf
}
\examples{
rpm_to_rcf(800)
}
\author{
Alex Zabeo
}

